@@ -67,38 +69,42 @@
     "roboto-fontface": "^0.10.0",
     "rxjs": "^6.6.7",
     "seedrandom": "^3.0.5",
-    "tslib": "^2.4.0",
+    "tslib": "^2.6.2",
     "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.12",
+    "@angular-devkit/build-angular": "^14.2.13",
     "@angular-eslint/builder": "^14.4.0",
     "@angular-eslint/eslint-plugin": "^14.4.0",
     "@angular-eslint/eslint-plugin-template": "^14.4.0",
     "@angular-eslint/schematics": "14.4.0",
     "@angular-eslint/template-parser": "^14.4.0",
-    "@angular/cli": "^14.2.12",
+    "@angular/cli": "^14.2.13",
     "@angular/compiler-cli": "^14.3.0",
-    "@babel/compat-data": "^7.22.5",
+    "@babel/compat-data": "^7.23.2",
     "@ngrx/schematics": "~14.3.3",
-    "@types/d3-selection": "^3.0.3",
-    "@types/hammerjs": "^2.0.41",
+    "@types/d3-selection": "^3.0.8",
+    "@types/hammerjs": "^2.0.43",
     "@types/jest": "^28.1.8",
+    "@types/lodash": "^4.14.200",
     "@types/node": "^12.20.55",
     "@types/ol": "^5.3.8",
-    "@types/seedrandom": "^3.0.2",
+    "@types/seedrandom": "^3.0.7",
     "@typescript-eslint/eslint-plugin": "5.43.0",
     "@typescript-eslint/parser": "5.43.0",
-    "eslint": "^8.27.0",
-    "eslint-plugin-rxjs": "~5.0.2",
+    "better-npm-audit": "^3.7.3",
+    "eslint": "^8.51.0",
+    "eslint-plugin-rxjs": "~5.0.3",
+    "glob": "^10.3.10",
     "jest": "^28.1.3",
-    "jest-canvas-mock": "^2.4.0",
+    "jest-canvas-mock": "^2.5.2",
     "jest-preset-angular": "^12.2.6",
     "rimraf": "^3.0.2",
     "rxjs-tslint": "^0.1.8",
     "rxjs-tslint-rules": "^4.34.8",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^20.0.0",
+    "stylelint": "^15.11.0",
+    "stylelint-config-standard": "^34.0.0",
+    "stylelint-config-standard-scss": "^11.0.0",
     "ts-node": "^8.10.2",
     "tsutils": "^3.21.0",
     "typescript": "4.6.4"
