@@ -98,8 +98,7 @@ export class FilterElementsViewComponent implements OnChanges {
         return !!this.inputData;
     }
 
-    // eslint-disable-next-line @typescript-eslint/ban-types
-    get filterTableResizeFlag(): {} {
+    get filterTableResizeFlag(): Record<string, never> {
         return this.filterTableResizeFlag_;
     }
 
@@ -113,15 +112,14 @@ export class FilterElementsViewComponent implements OnChanges {
     moreFilterOpenState = false;
     complexFilterOpenState = false;
 
-    // eslint-disable-next-line @typescript-eslint/ban-types
-    private filterTableResizeFlag_: {} = {};
+    private filterTableResizeFlag_: Record<string, never> = {};
 
     private checkTableSizeSubject_ = new Subject<number>();
     checkTableSize$ = this.checkTableSizeSubject_.asObservable();
     private updateTableSizeSubject_ = new Subject<void>();
     updateTableSize$ = this.updateTableSizeSubject_.asObservable();
 
-    // lifecycle hooks start
+    //#region lifecycle hooks
 
     ngOnChanges(changes: SimpleChanges): void {
         if (changes.inputData !== undefined && changes.inputData.currentValue !== null) {
@@ -129,9 +127,9 @@ export class FilterElementsViewComponent implements OnChanges {
         }
     }
 
-    // life cycle hooks end
+    //#endregion life cycle hooks
 
-    // template triggers start
+    //#region template triggers
 
     onClearAllFilters(): void {
         this.clearAllFilters.emit();
@@ -196,7 +194,7 @@ export class FilterElementsViewComponent implements OnChanges {
         this.updateTableSizeSubject_.next();
     }
 
-    // template triggers end
+    //#endregion template triggers
 
     private processInputData(): void {
         if (
@@ -211,12 +209,16 @@ export class FilterElementsViewComponent implements OnChanges {
         ) {
             this.checkTableSizeSubject_.next(200);
         }
-        this.updateFilterAndRows();
         this.updateDataColumns();
+        this.updateFilterAndRows();
         this.updateTableInputData();
         this.updatePropValueMap();
 
         this.processedInput__ = this.inputData;
+    }
+
+    private getIgnoredProps(): string[] {
+        return this.dataColumns_.filter(c => c.unavailable).map(c => c.id);
     }
 
     private updateFilterAndRows(): void {
@@ -231,8 +233,16 @@ export class FilterElementsViewComponent implements OnChanges {
                     oldFilterMap.predefinedFilter
             ),
             complexFilter: (
-                !oldSettings || oldSettings.complexFilter !== newSettings.complexFilter ?
-                    getUpdatedComplexRowFilter(newSettings.complexFilter, oldFilterMap ? oldFilterMap.complexFilter : undefined) :
+                (
+                    !oldSettings ||
+                    oldSettings.complexFilter !== newSettings.complexFilter ||
+                    this.processedInput__.dataTable.columns !== this.inputData.dataTable.columns
+                ) ?
+                    getUpdatedComplexRowFilter(
+                        newSettings.complexFilter,
+                        this.getIgnoredProps(),
+                        oldFilterMap ? oldFilterMap.complexFilter : undefined
+                    ) :
                     oldFilterMap.complexFilter
             ),
             standardFilter: (
@@ -270,46 +280,22 @@ export class FilterElementsViewComponent implements OnChanges {
     }
 
     private updateTableInputData(): void {
-        if (!this.filterTableViewInputData_) {
+        if (
+            !this.filterTableViewInputData_ ||
+            this.filterTableViewInputData_.dataTable !== this.inputData.dataTable ||
+            this.filterTableViewInputData_.filteredRows !== this.prefilteredRows_ ||
+            this.filterTableViewInputData_.visibilityFilter !== this.inputData.filterTableSettings.visibilityFilter ||
+            this.filterTableViewInputData_.columnFilters !== this.inputData.filterTableSettings.columnFilters ||
+            this.filterTableViewInputData_.columnOrder !== this.inputData.filterTableSettings.columnOrder
+        ) {
             this.filterTableViewInputData_ = {
-                dataTable: {
-                    ...this.inputData.dataTable,
-                    rows: this.prefilteredRows_
-                },
+                dataTable: this.inputData.dataTable,
+                filteredRows: this.prefilteredRows_,
                 columnOrder: this.inputData.filterTableSettings.columnOrder,
                 selectedRowIds: this.inputData.selectedRowIds,
                 visibilityFilter: this.inputData.filterTableSettings.visibilityFilter,
                 columnFilters: this.inputData.filterTableSettings.columnFilters
             };
-        } else {
-            const dataTable = (
-                (
-                    this.inputData.dataTable.columns !== this.filterTableViewInputData_.dataTable.columns ||
-                    this.prefilteredRows_ !== this.filterTableViewInputData_.dataTable.rows
-                ) ?
-                    {
-                        ...this.inputData.dataTable,
-                        rows: this.prefilteredRows_
-                    } :
-                    this.filterTableViewInputData_.dataTable
-            );
-
-            this.filterTableViewInputData_ = (
-                (
-                    this.filterTableViewInputData_.dataTable !== dataTable ||
-                    this.filterTableViewInputData_.visibilityFilter !== this.inputData.filterTableSettings.visibilityFilter ||
-                    this.filterTableViewInputData_.columnFilters !== this.inputData.filterTableSettings.columnFilters ||
-                    this.filterTableViewInputData_.columnOrder !== this.inputData.filterTableSettings.columnOrder
-                ) ?
-                    ({
-                        dataTable: dataTable,
-                        columnOrder: this.inputData.filterTableSettings.columnOrder,
-                        selectedRowIds: this.inputData.selectedRowIds,
-                        visibilityFilter: this.inputData.filterTableSettings.visibilityFilter,
-                        columnFilters: this.inputData.filterTableSettings.columnFilters
-                    }) :
-                    this.filterTableViewInputData_
-            );
         }
     }
 
