@@ -1,24 +1,34 @@
 import {
-    DeliveryStoreData as DeliveryData, FclData, ObservedType, StationStoreData as StationData,
+    DeliveryStoreData as DeliveryData,
+    FclData, ObservedType,
+    StationStoreData as StationData,
     Connection, GroupType, GroupData,
     ValueCondition as IntValueCondition,
     LogicalCondition as IntLogicalCondition,
-    ValueType, OperationType, NodeShapeType, LinePatternType, MergeDeliveriesType, CrossContTraceType, StationId, DeliveryId
+    ValueType, OperationType, NodeShapeType, LinePatternType,
+    MergeDeliveriesType, CrossContTraceType, StationId, DeliveryId, HighlightingRule,
+    LabelPart as IntLabelPart,
+    StationHighlightingRule as IntStationHighlightingRule
 } from '../../data.model';
 import { HttpClient } from '@angular/common/http';
 
 import { Utils } from '../../util/non-ui-utils';
 import * as ExtDataConstants from '../ext-data-constants.v1';
-
 import { IDataImporter } from './datatypes';
-import { isValidJson, createDefaultHighlights, checkVersionFormat, areMajorVersionsMatching } from './shared';
+import {
+    isValidJson, checkVersionFormat,
+    areMajorVersionsMatching,
+    createDefaultStationAnonymizationLabelHRule,
+    createDefaultStationHRules,
+    createDefaultDeliveryHRules
+} from './shared';
 import { importSamples } from './sample-importer-v1';
 import {
     ViewData,
-    StationHighlightingRule as ExtStationHighlightingRule,
     DeliveryHighlightingRule as ExtDeliveryHighlightingRule,
     ValueCondition as ExtValueCondition,
     LogicalCondition as ExtLogicalCondition,
+    AnonymizationRule as ExtAnonymizationRule,
     JsonData,
     VERSION,
     MetaNodeData
@@ -422,9 +432,8 @@ export class DataImporterV1 implements IDataImporter {
             jsonData.settings.view === undefined ||
             jsonData.settings.view === null
         ) {
-            const defaultHS = createDefaultHighlights();
-            fclData.graphSettings.highlightingSettings.stations = defaultHS.stations;
-            fclData.graphSettings.highlightingSettings.deliveries = defaultHS.deliveries;
+            fclData.graphSettings.highlightingSettings.stations = createDefaultStationHRules(true);
+            fclData.graphSettings.highlightingSettings.deliveries = createDefaultDeliveryHRules();
             return;
         }
 
@@ -530,32 +539,46 @@ export class DataImporterV1 implements IDataImporter {
     }
 
     private convertExternalHighlightingSettings(viewData: ViewData, fclData: FclData): void {
-        if (viewData && viewData.node && viewData.node.highlightConditions) {
+        const extStatHighlightingRules = viewData?.node?.highlightConditions || undefined;
+        const extStatAnoRule = viewData?.node?.anonymizationRule || undefined;
+        const extToIntStatPropMap = this.createReverseMapFromSimpleMap(fclData.source.propMaps.stationPropMap);
 
-            const extStatHighlightingRules: ExtStationHighlightingRule[] = viewData.node.highlightConditions;
+        if (extStatHighlightingRules) {
 
             if (extStatHighlightingRules.length > 0) {
-                const extToIntPropMap = this.createReverseMapFromSimpleMap(fclData.source.propMaps.stationPropMap);
 
                 fclData.graphSettings.highlightingSettings.stations = extStatHighlightingRules.map((extRule, extRuleIndex) => (
                     {
                         id: 'SHR' + extRuleIndex,
                         name: extRule.name,
                         showInLegend: extRule.showInLegend === true,
-                        disabled: extRule.disabled === true,
+                        userDisabled: extRule.disabled === true,
+                        autoDisabled: false,
                         color: extRule.color,
                         invisible: extRule.invisible,
                         adjustThickness: extRule.adjustThickness,
-                        labelProperty: this.mapLabelProperty(extRule.labelProperty, extToIntPropMap),
-                        valueCondition: this.mapValueCondition(extRule.valueCondition, extToIntPropMap),
-                        logicalConditions: this.mapLogicalConditions(extRule.logicalConditions, extToIntPropMap),
+                        labelProperty: this.mapLabelProperty(extRule.labelProperty, extToIntStatPropMap),
+                        valueCondition: this.mapValueCondition(extRule.valueCondition, extToIntStatPropMap),
+                        logicalConditions: this.mapLogicalConditions(extRule.logicalConditions, extToIntStatPropMap),
                         shape: this.mapShapeType(extRule.shape)
                     }
                 ));
             }
+
         } else {
-            fclData.graphSettings.highlightingSettings.stations = createDefaultHighlights().stations;
+            fclData.graphSettings.highlightingSettings.stations = createDefaultStationHRules(false);
         }
+
+        const intStatAnoHRule: IntStationHighlightingRule = extStatAnoRule ?
+            { ...this.convertExternalAnoRule(extStatAnoRule, extToIntStatPropMap), shape: null } :
+            createDefaultStationAnonymizationLabelHRule();
+
+        if (intStatAnoHRule.userDisabled === false) {
+            fclData.graphSettings.highlightingSettings.stations.forEach(r => r.autoDisabled = true);
+        }
+
+        fclData.graphSettings.highlightingSettings.stations.push(intStatAnoHRule);
+
 
         if (viewData && viewData.edge && viewData.edge.highlightConditions) {
 
@@ -569,7 +592,9 @@ export class DataImporterV1 implements IDataImporter {
                         id: 'DHR' + extRuleIndex,
                         name: extRule.name,
                         showInLegend: extRule.showInLegend === true,
-                        disabled: extRule.disabled === true,
+                        userDisabled: extRule.disabled === true,
+                        autoDisabled: false,
+                        activationDisablesOtherRules: false,
                         color: extRule.color,
                         invisible: extRule.invisible,
                         adjustThickness: extRule.adjustThickness,
@@ -581,8 +606,38 @@ export class DataImporterV1 implements IDataImporter {
                 ));
             }
         } else {
-            fclData.graphSettings.highlightingSettings.deliveries = createDefaultHighlights().deliveries;
+            fclData.graphSettings.highlightingSettings.deliveries = createDefaultDeliveryHRules();
         }
+    }
+
+    private convertExternalAnoRule(extAnoRule: ExtAnonymizationRule, extToIntPropMap: Map<string, string>): HighlightingRule {
+        const defaultIntAnoHRule = createDefaultStationAnonymizationLabelHRule();
+        let labelParts = extAnoRule.labelParts.map(
+            p => (
+                p.property ?
+                    { prefix: p.prefix, property: extToIntPropMap.get(p.property) } :
+                    { prefix: p.prefix, useIndex: p.useIndex || false }
+            ) as IntLabelPart
+        );
+        const indexParts = labelParts.filter(p => p.useIndex !== undefined);
+        if (indexParts.length === 0) {
+            labelParts.push(...defaultIntAnoHRule.labelParts.filter(p => p.useIndex !== undefined));
+        } else if (indexParts.length > 1) {
+            labelParts = _.difference(labelParts, indexParts.slice(1));
+        }
+
+        const intAnoHRule: HighlightingRule = {
+            ...defaultIntAnoHRule,
+            labelPrefix: extAnoRule.labelPrefix,
+            userDisabled: extAnoRule.disabled,
+            labelParts: extAnoRule.labelParts.map(
+                p => p.property ?
+                    { prefix: p.prefix, property: extToIntPropMap.get(p.property) } :
+                    { prefix: p.prefix, useIndex: p.useIndex || false }
+            ),
+            logicalConditions: this.mapLogicalConditions(extAnoRule.logicalConditions, extToIntPropMap)
+        };
+        return intAnoHRule;
     }
 
     private mapLogicalConditions(
