@@ -1,12 +1,23 @@
 
-import { GroupType, OperationType, ValueType, NodeShapeType, MergeDeliveriesType, Connection,
-    StationStoreData, DeliveryStoreData } from '../../data.model';
+import {
+    GroupType, OperationType, ValueType,
+    NodeShapeType, MergeDeliveriesType, Connection,
+    StationStoreData, DeliveryStoreData
+} from '../../data.model';
 import { Map as ImmutableMap, List as ImmutableList } from 'immutable';
 import * as ExtDataConstants from './../ext-data-constants.v1';
 import { Constants as IntDataConstants } from '../../util/constants';
-import { DataTable, DataRow, JsonData, ColumnProperty, HighlightingRule as ExtHighlightingRule } from '../ext-data-model.v1';
+import {
+    DataTable, DataRow, JsonData, ColumnProperty,
+    HighlightingRule as ExtHighlightingRule,
+    LogicalCondition as ExtLogicalCondition
+} from '../ext-data-model.v1';
 import { Utils } from '../../util/non-ui-utils';
-import { STATION_PROP_TO_REQ_TYPE_MAP, DELIVERY_PROP_TO_REQ_TYPE_MAP, DEL2DEL_PROP_TO_REQ_TYPE_MAP } from '../int-data-constants';
+import {
+    STATION_PROP_TO_REQ_TYPE_MAP,
+    DELIVERY_PROP_TO_REQ_TYPE_MAP,
+    DEL2DEL_PROP_TO_REQ_TYPE_MAP
+} from '../int-data-constants';
 import { isValueTypeValid } from './shared';
 import * as _ from 'lodash';
 import { InputDataError } from '../io-errors';
@@ -248,7 +259,7 @@ function getReferencedProps(highlightingConditions: ExtHighlightingRule[]): stri
                     hCon.labelProperty,
                     ...(
                         hCon.logicalConditions ?
-                            [].concat(...hCon.logicalConditions.map(logConA => logConA.map(logCon => logCon.propertyName))) :
+                            getLogicalConditionsProps(hCon.logicalConditions) :
                             []
                     ),
                     ...hCon.valueCondition ? [hCon.valueCondition.propertyName] : []
@@ -258,19 +269,36 @@ function getReferencedProps(highlightingConditions: ExtHighlightingRule[]): stri
     );
 }
 
+function getLogicalConditionsProps(logicalConditions: ExtLogicalCondition[][]): string[] {
+    return [].concat(...logicalConditions.map(logConA => logConA.map(logCon => logCon.propertyName)));
+}
+
 export function getStationPropMap(jsonData: JsonData): PropMap {
+    let referencedProps: string[] = [];
+    if (
+        jsonData.settings &&
+        jsonData.settings.view &&
+        jsonData.settings.view.node) {
+
+        if (jsonData.settings.view.node.highlightConditions) {
+            referencedProps = getReferencedProps(jsonData.settings.view.node.highlightConditions);
+        }
+
+        const anoRule = jsonData.settings.view.node.anonymizationRule;
+        if (anoRule) {
+            const logicalConditionProps = getLogicalConditionsProps(anoRule.logicalConditions);
+            const labelPartProps = anoRule.labelParts.map(p => p.property).filter(p => p !== undefined);
+            referencedProps = [].concat(
+                referencedProps,
+                logicalConditionProps,
+                labelPartProps
+            );
+        }
+    }
+
     return getPropMap(
         jsonData.data.stations,
-        getReferencedProps(
-            (
-                jsonData.settings &&
-                jsonData.settings.view &&
-                jsonData.settings.view.node &&
-                jsonData.settings.view.node.highlightConditions
-            ) ?
-                jsonData.settings.view.node.highlightConditions :
-                []
-        ),
+        referencedProps,
         DEFAULT_STATION_PROP_INT_TO_EXT_MAP,
         STATION_PROPS_INT_TO_EXT_ALT_MAP,
         IntDataConstants.STATION_PROPERTIES.toArray()
