@@ -1,4 +1,4 @@
-export const VERSION = '1.1.2';
+export const VERSION = '1.2.0';
 
 export interface JsonData {
     version: string;
@@ -72,12 +72,26 @@ export interface ViewData {
     // explosions?: ExplosionViewData[];
 }
 
+export interface LabelPart {
+    prefix: string;
+    property?: string;
+    useIndex?: boolean;
+}
+
+export interface AnonymizationRule {
+    labelParts: LabelPart[];
+    labelPrefix: string;
+    disabled: boolean;
+    logicalConditions: LogicalCondition[][];
+}
+
 export interface NodeViewData {
     skipEdgelessNodes?: boolean;
     // labelPosition: string;
     selectedNodes?: string[];
     invisibleNodes?: string[] | null;
     highlightConditions?: StationHighlightingRule[];
+    anonymizationRule?: AnonymizationRule | null;
 }
 
 export interface EdgeViewData {
